Workaround for Performance issue in 3.5 due to query splitting
The query splitting is triggered during execution of an extra rule in the updated Spark 3.5's RuleExecutor (named V2ScanPartitioningAndOrdering) which is not there in Spark 3.1 and hence is not triggered before it needs to be. This causes a lot of network calls and gives time lag.
The workaround for now is not to trigger Splitting when the function outputPartitioning (this is the function in the extra rule that triggers splitting) by setting its value to the default spark value - UnknownPartitioning.

Might need readdressing in future when we start making use of partition other than UnknownPartitioning.